Develop

+++ app/src/main/java/mx/kodemia/bankodemiaapp/data/model/request/MakeTransactionRequest.kt
@@ -1,9 +1,7 @@
 package mx.kodemia.bankodemiaapp.data.model.request
 
-import mx.kodemia.bankodemiaapp.data.model.response.makeTransaction.Type
-
 data class MakeTransactionRequest(
-    val amount: Double,
+    val amount: Int,
     val type: String,
     val destinationUser: String?,
     val concept: String
